refactor: time util functions
